refactor: 💡 Modal structure refactored and integration made
Changed the modal structure thinking about the project's scalability,
with the refactoring it can receive any other component internally.
ModalProductContent component created and integration made. Changed
ISales interface to ISale, i forgot to commit this file in the last
commit and the build on Vercel failed.
